fixed output of coffee and primary banner
